Extract BandWidget to a separate module
